pokemon description fragment logic added

+++ app/src/main/java/ru/eugeneproj/pokemon/data/model/pokemoninfomodel/PokemonInfo.kt
@@ -1,10 +1,10 @@
 package ru.eugeneproj.pokemon.data.model.pokemoninfomodel
 
 data class PokemonInfo(
-    val height: Int,
+    val height: Float,
     val name: String,
     val sprites: Sprites,
     val stats: List<Stat>,
     val types: List<Type>,
-    val weight: Int
+    val weight: Float
 )
